+++ internal/controller/promotions/argocd_test.go
@@ -1,11 +1,10 @@
-package controller
+package promotions
 
 import (
 	"context"
 	"testing"
 
 	argocd "github.com/argoproj/argo-cd/v2/pkg/apis/application/v1alpha1"
-	argoHealth "github.com/argoproj/gitops-engine/pkg/health"
 	"github.com/pkg/errors"
 	"github.com/stretchr/testify/require"
 	v1 "k8s.io/apimachinery/pkg/apis/meta/v1"
@@ -13,256 +12,6 @@ import (
 
 	api "github.com/akuityio/kargo/api/v1alpha1"
 )
-
-func TestCheckHealth(t *testing.T) {
-	testCases := []struct {
-		name           string
-		state          api.EnvironmentState
-		appChecks      []api.ArgoCDAppCheck
-		getArgoCDAppFn func(
-			context.Context,
-			client.Client,
-			string,
-			string,
-		) (*argocd.Application, error)
-		assertions func(api.Health)
-	}{
-		{
-			name: "healthchecks do not include any Argo CD Apps",
-			assertions: func(health api.Health) {
-				require.Equal(t,
-					api.Health{
-						Status: api.HealthStateUnknown,
-						StatusReason: "spec.healthChecks contains insufficient " +
-							"instructions to assess Environment health",
-					},
-					health,
-				)
-			},
-		},
-
-		{
-			name: "error finding Argo CD App",
-			appChecks: []api.ArgoCDAppCheck{
-				{
-					AppName:      "fake-app",
-					AppNamespace: "fake-namespace",
-				},
-			},
-			getArgoCDAppFn: func(
-				context.Context,
-				client.Client,
-				string,
-				string,
-			) (*argocd.Application, error) {
-				return nil, errors.New("something went wrong")
-			},
-			assertions: func(health api.Health) {
-				require.Equal(t, api.HealthStateUnknown, health.Status)
-				require.Contains(
-					t,
-					health.StatusReason,
-					"error finding Argo CD Application",
-				)
-				require.Contains(t, health.StatusReason, "something went wrong")
-			},
-		},
-
-		{
-			name: "Argo CD App not found",
-			appChecks: []api.ArgoCDAppCheck{
-				{
-					AppName:      "fake-app",
-					AppNamespace: "fake-namespace",
-				},
-			},
-			getArgoCDAppFn: func(
-				context.Context,
-				client.Client,
-				string,
-				string,
-			) (*argocd.Application, error) {
-				return nil, nil
-			},
-			assertions: func(health api.Health) {
-				require.Equal(t, api.HealthStateUnknown, health.Status)
-				require.Contains(
-					t,
-					health.StatusReason,
-					"unable to find Argo CD Application",
-				)
-			},
-		},
-
-		{
-			name: "Argo CD App is multi-source",
-			// This doesn't require there to actually BE multiple sources. Simply
-			// using the sources field instead of the source fields should be enough
-			// to trigger this case.
-			appChecks: []api.ArgoCDAppCheck{
-				{
-					AppName:      "fake-app",
-					AppNamespace: "fake-namespace",
-				},
-			},
-			getArgoCDAppFn: func(
-				context.Context,
-				client.Client,
-				string,
-				string,
-			) (*argocd.Application, error) {
-				return &argocd.Application{
-					Spec: argocd.ApplicationSpec{
-						Sources: argocd.ApplicationSources{
-							{},
-						},
-					},
-				}, nil
-			},
-			assertions: func(health api.Health) {
-				require.Equal(t, api.HealthStateUnknown, health.Status)
-				require.Contains(
-					t,
-					health.StatusReason,
-					"bugs in Argo CD currently prevent a comprehensive assessment of "+
-						"the health of multi-source Application",
-				)
-			},
-		},
-
-		{
-			name: "Argo CD App is not healthy",
-			appChecks: []api.ArgoCDAppCheck{
-				{
-					AppName:      "fake-app",
-					AppNamespace: "fake-namespace",
-				},
-			},
-			getArgoCDAppFn: func(
-				context.Context,
-				client.Client,
-				string,
-				string,
-			) (*argocd.Application, error) {
-				return &argocd.Application{
-					Status: argocd.ApplicationStatus{
-						Health: argocd.HealthStatus{
-							Status: argoHealth.HealthStatusDegraded,
-						},
-					},
-				}, nil
-			},
-			assertions: func(health api.Health) {
-				require.Equal(t, api.HealthStateUnhealthy, health.Status)
-				require.Contains(t, health.StatusReason, "has health state")
-				require.Contains(
-					t,
-					health.StatusReason,
-					argoHealth.HealthStatusDegraded,
-				)
-			},
-		},
-
-		{
-			name: "Argo CD App not synced",
-			appChecks: []api.ArgoCDAppCheck{
-				{
-					AppName:      "fake-app",
-					AppNamespace: "fake-namespace",
-				},
-			},
-			getArgoCDAppFn: func(
-				context.Context,
-				client.Client,
-				string,
-				string,
-			) (*argocd.Application, error) {
-				return &argocd.Application{
-					Spec: argocd.ApplicationSpec{
-						Source: &argocd.ApplicationSource{},
-					},
-					Status: argocd.ApplicationStatus{
-						Health: argocd.HealthStatus{
-							Status: argoHealth.HealthStatusHealthy,
-						},
-						Sync: argocd.SyncStatus{
-							Status: argocd.SyncStatusCodeOutOfSync,
-						},
-					},
-				}, nil
-			},
-			assertions: func(health api.Health) {
-				require.Equal(t, api.HealthStateUnhealthy, health.Status)
-				require.Contains(
-					t,
-					health.StatusReason,
-					"is not synced to revision",
-				)
-			},
-		},
-
-		{
-			name: "Argo CD App healthy and synced",
-			state: api.EnvironmentState{
-				Commits: []api.GitCommit{
-					{
-						RepoURL: "fake-url",
-						ID:      "fake-commit",
-					},
-				},
-			},
-			appChecks: []api.ArgoCDAppCheck{
-				{
-					AppName:      "fake-app",
-					AppNamespace: "fake-namespace",
-				},
-			},
-			getArgoCDAppFn: func(
-				context.Context,
-				client.Client,
-				string,
-				string,
-			) (*argocd.Application, error) {
-				return &argocd.Application{
-					Spec: argocd.ApplicationSpec{
-						Source: &argocd.ApplicationSource{
-							RepoURL: "fake-url",
-						},
-					},
-					Status: argocd.ApplicationStatus{
-						Health: argocd.HealthStatus{
-							Status: argoHealth.HealthStatusHealthy,
-						},
-						Sync: argocd.SyncStatus{
-							Status:   argocd.SyncStatusCodeSynced,
-							Revision: "fake-commit",
-						},
-					},
-				}, nil
-			},
-			assertions: func(health api.Health) {
-				require.Equal(t, api.HealthStateHealthy, health.Status)
-				require.Empty(t, health.StatusReason)
-			},
-		},
-	}
-	for _, testCase := range testCases {
-		t.Run(testCase.name, func(t *testing.T) {
-			reconciler := &environmentReconciler{
-				getArgoCDAppFn: testCase.getArgoCDAppFn,
-			}
-			testCase.assertions(
-				reconciler.checkHealth(
-					context.Background(),
-					testCase.state,
-					api.HealthChecks{
-						ArgoCDAppChecks: testCase.appChecks,
-					},
-				),
-			)
-		})
-	}
-}
 
 func TestApplyArgoCDSourceUpdate(t *testing.T) {
 	testCases := []struct {
@@ -464,7 +213,7 @@ func TestApplyArgoCDSourceUpdate(t *testing.T) {
 			},
 		},
 	}
-	reconciler := &promotionReconciler{}
+	reconciler := &reconciler{}
 	for _, testCase := range testCases {
 		t.Run(testCase.name, func(t *testing.T) {
 			updatedSource, err := reconciler.applyArgoCDSourceUpdate(
@@ -527,7 +276,7 @@ func TestAuthorizeArgoCDAppUpdate(t *testing.T) {
 			allowed: true,
 		},
 	}
-	reconciler := promotionReconciler{}
+	reconciler := reconciler{}
 	for _, testCase := range testCases {
 		t.Run(testCase.name, func(t *testing.T) {
 			err := reconciler.authorizeArgoCDAppUpdate(
@@ -762,7 +511,7 @@ func TestApplyArgoCDAppUpdate(t *testing.T) {
 	}
 	for _, testCase := range testCases {
 		t.Run(testCase.name, func(t *testing.T) {
-			reconciler := promotionReconciler{
+			reconciler := reconciler{
 				getArgoCDAppFn:            testCase.getArgoCDAppFn,
 				applyArgoCDSourceUpdateFn: testCase.applyArgoCDSourceUpdateFn,
 				patchFn:                   testCase.patchFn,
