+++ internal/controller/promotions/git_test.go
@@ -1,4 +1,4 @@
-package controller
+package promotions
 
 import (
 	"context"
@@ -26,8 +26,8 @@ func TestApplyGitRepoUpdate(t *testing.T) {
 	}{
 		{
 			name: "error getting repo credentials",
-			credentialsDB: &fakeCredentialsDB{
-				getFn: func(
+			credentialsDB: &credentials.FakeDB{
+				GetFn: func(
 					context.Context,
 					string,
 					credentials.Type,
@@ -50,8 +50,8 @@ func TestApplyGitRepoUpdate(t *testing.T) {
 
 		{
 			name: "error applying updates",
-			credentialsDB: &fakeCredentialsDB{
-				getFn: func(
+			credentialsDB: &credentials.FakeDB{
+				GetFn: func(
 					context.Context,
 					string,
 					credentials.Type,
@@ -76,8 +76,8 @@ func TestApplyGitRepoUpdate(t *testing.T) {
 
 		{
 			name: "success",
-			credentialsDB: &fakeCredentialsDB{
-				getFn: func(
+			credentialsDB: &credentials.FakeDB{
+				GetFn: func(
 					context.Context,
 					string,
 					credentials.Type,
@@ -107,7 +107,7 @@ func TestApplyGitRepoUpdate(t *testing.T) {
 	}
 	for _, testCase := range testCases {
 		t.Run(testCase.name, func(t *testing.T) {
-			reconciler := promotionReconciler{
+			reconciler := reconciler{
 				credentialsDB:    testCase.credentialsDB,
 				gitApplyUpdateFn: testCase.gitApplyUpdateFn,
 			}
@@ -128,127 +128,6 @@ func TestApplyGitRepoUpdate(t *testing.T) {
 				},
 			)
 			testCase.assertions(newState, outState, err)
-		})
-	}
-}
-
-func TestGetLatestCommits(t *testing.T) {
-	testCases := []struct {
-		name                string
-		credentialsDB       credentials.Database
-		getLatestCommitIDFn func(
-			string,
-			string,
-			*git.Credentials,
-		) (string, error)
-		assertions func(commits []api.GitCommit, err error)
-	}{
-		{
-			name: "error getting repo credentials",
-			credentialsDB: &fakeCredentialsDB{
-				getFn: func(
-					context.Context,
-					string,
-					credentials.Type,
-					string,
-				) (credentials.Credentials, bool, error) {
-					return credentials.Credentials{}, false,
-						errors.New("something went wrong")
-				},
-			},
-			assertions: func(commits []api.GitCommit, err error) {
-				require.Error(t, err)
-				require.Contains(
-					t,
-					err.Error(),
-					"error obtaining credentials for git repo",
-				)
-				require.Contains(t, err.Error(), "something went wrong")
-				require.Empty(t, commits)
-			},
-		},
-
-		{
-			name: "error getting last commit ID",
-			credentialsDB: &fakeCredentialsDB{
-				getFn: func(
-					context.Context,
-					string,
-					credentials.Type,
-					string,
-				) (credentials.Credentials, bool, error) {
-					return credentials.Credentials{}, false, nil
-				},
-			},
-			getLatestCommitIDFn: func(
-				string,
-				string,
-				*git.Credentials,
-			) (string, error) {
-				return "", errors.New("something went wrong")
-			},
-			assertions: func(commits []api.GitCommit, err error) {
-				require.Error(t, err)
-				require.Contains(
-					t,
-					err.Error(),
-					"error determining latest commit ID of git repo",
-				)
-				require.Contains(t, err.Error(), "something went wrong")
-				require.Empty(t, commits)
-			},
-		},
-
-		{
-			name: "success",
-			credentialsDB: &fakeCredentialsDB{
-				getFn: func(
-					context.Context,
-					string,
-					credentials.Type,
-					string,
-				) (credentials.Credentials, bool, error) {
-					return credentials.Credentials{}, false, nil
-				},
-			},
-			getLatestCommitIDFn: func(
-				string,
-				string,
-				*git.Credentials,
-			) (string, error) {
-				return "fake-commit", nil
-			},
-			assertions: func(commits []api.GitCommit, err error) {
-				require.NoError(t, err)
-				require.Len(t, commits, 1)
-				require.Equal(
-					t,
-					api.GitCommit{
-						RepoURL: "fake-url",
-						ID:      "fake-commit",
-					},
-					commits[0],
-				)
-			},
-		},
-	}
-	for _, testCase := range testCases {
-		t.Run(testCase.name, func(t *testing.T) {
-			reconciler := environmentReconciler{
-				credentialsDB:       testCase.credentialsDB,
-				getLatestCommitIDFn: testCase.getLatestCommitIDFn,
-			}
-			testCase.assertions(
-				reconciler.getLatestCommits(
-					context.Background(),
-					"fake-namespace",
-					[]api.GitSubscription{
-						{
-							RepoURL: "fake-url",
-						},
-					},
-				),
-			)
 		})
 	}
 }
